+++ src/Setup/BlockPatterns.php
@@ -33,10 +33,22 @@ class BlockPatterns implements SetupInterface
     private $remove_patterns = [];
 
     /**
+     * Remove patterns by prefix
+     * @var array
+     */
+    private $remove_patterns_by_prefix = [];
+
+    /**
      * Categories
      * @var array
      */
     private $categories = [];
+
+    /**
+     * Remove categories
+     * @var array
+     */
+    private $remove_categories = [];
 
     /**
      * Construct class
@@ -73,12 +85,32 @@ class BlockPatterns implements SetupInterface
 
     /**
      * Remove pattern
-     * @param string $name Pattern name.
+     * @param string $slug Pattern slug.
      * @return void
      */
-    public function removePattern(string $name)
+    public function removePattern(string $slug)
     {
-        $this->remove_patterns[] = $name;
+        $this->remove_patterns[] = $slug;
+    }
+
+    /**
+     * Remove pattern by prefix
+     * @param string $prefix Pattern prefix.
+     * @return void
+     */
+    public function removePatternByPrefix(string $prefix)
+    {
+        $this->remove_patterns_by_prefix[] = $prefix;
+    }
+
+    /**
+     * Remove category
+     * @param string $name Category name.
+     * @return void
+     */
+    public function removeCategory(string $name)
+    {
+        $this->remove_categories[] = $name;
     }
 
     /**
@@ -87,22 +119,32 @@ class BlockPatterns implements SetupInterface
      */
     public function init()
     {
-        if (!empty($this->patterns) || !empty($this->categories) || !empty($this->remove_patterns)) {
+        if (!empty($this->patterns) || !empty($this->categories) || !empty($this->remove_patterns) || !empty($this->remove_categories)) {
             add_action('init', function () {
 
                 if (function_exists('register_block_pattern_category')) {
+                    
                     foreach ($this->categories as $category) {
                         register_block_pattern_category(
                             \sanitize_title($category),
                             array('label' => $category)
                         );
                     }
-                    if (!empty($this->remove_patterns)) {
+                    
+                    foreach ($this->remove_categories as $category) {
+                        unregister_block_pattern_category(
+                            \sanitize_title($category)
+                        );
+                    }
+
+                    if (!empty($this->remove_patterns)||!empty($this->remove_patterns_by_prefix)) {
 
                         $registered_patterns = \WP_Block_Patterns_Registry::get_instance()->get_all_registered();
                         if ( $registered_patterns ) {
                             foreach ( $registered_patterns as $pattern_properties ) {
-                                if ( in_array( $pattern_properties['name'], $this->remove_patterns )) {
+                                if ( in_array( substr( $pattern_properties['slug'], 0, strpos( $pattern_properties['slug'], '/' ) ), $this->remove_patterns_by_prefix ) ) {
+                                    unregister_block_pattern( $pattern_properties['name'] );
+                                } elseif ( in_array( $pattern_properties['name'], $this->remove_patterns )) {
                                     unregister_block_pattern( $pattern_properties['name'] );
                                 }
                             }